support resume_all
